refactor: split wandb staging and payload assembly
